refactor(core): split the numeric record batch out, and say "field" where a caller reads
Three changes, no behavior difference.

`NumericRecordBatch` moves to its own module, as `NumericRecord` sits beside
`Record` for the same reason: the numeric specialization adds a flat-vector
contract of its own and nothing in the general case depends on it. That takes the
base module from 1143 lines to 919, under the threshold the style guide sets.

The contract prose now says *field* where a caller reads it and keeps *column*
for the storage layer, which is what the word describes. A caller addresses
fields -- `batch["x"]` hands back that field's values across the batch -- and the
constructor's first argument becomes `fields`, the name `Record` and `RecordArray`
already use in that position. The internal slot, its helpers, and "columnar" as a
description of the layout are unchanged.

Two single-caller helpers are inlined, two that serve the pytree hook move beside
it, and the two about object-array storage move to the module that owns that
storage, next to the array construction they belong with. Thirteen helpers where
there were sixteen, each now sitting with the concern it serves.
