+++ probpipe/core/_object_batch.py
@@ -156,6 +156,27 @@ class _ObjectBatch[E](Batch[E]):
         return view
 
 
+def _frozen_object_column(column: np.ndarray) -> np.ndarray:
+    """*column* as an object array nobody can write through.
+
+    A batch holds the columns it validated. An object column is the one kind a
+    caller can still mutate after construction — a JAX array is already immutable
+    and a numpy numeric column follows the aliasing convention the single-record
+    types already set — so it is copied and frozen, for the reason
+    ``_ObjectBatch`` states: a caller keeping a handle on what they passed cannot
+    write a value into the batch that its spec does not admit. Only the pointer
+    array is copied, so the elements stay shared.
+    """
+    frozen = np.array(column, dtype=object, subok=False)
+    frozen.setflags(write=False)
+    return frozen
+
+
+def _is_object_array(column: Any) -> bool:
+    """Whether *column* is a numpy array of objects, the non-array column form."""
+    return isinstance(column, np.ndarray) and column.dtype == object
+
+
 def _as_object_array(elements: np.ndarray | Iterable[Any], *, kind: str) -> np.ndarray:
     """*elements* as a writable-by-nobody object array, without unpacking it.
